Move create table statement to sql from GlobalStateMgr to AstToString
