+++ fe/fe-core/src/test/java/com/starrocks/analysis/ShowCreateMaterializedViewStmtTest.java
@@ -21,6 +21,7 @@ import com.starrocks.common.Config;
 import com.starrocks.qe.ConnectContext;
 import com.starrocks.server.GlobalStateMgr;
 import com.starrocks.server.MetadataMgr;
+import com.starrocks.sql.analyzer.AstToStringBuilder;
 import com.starrocks.sql.analyzer.SemanticException;
 import com.starrocks.sql.ast.CreateMaterializedViewStatement;
 import com.starrocks.sql.ast.StatementBase;
@@ -98,7 +99,7 @@ public class ShowCreateMaterializedViewStmtTest {
         currentState.createMaterializedView((CreateMaterializedViewStatement) statementBase);
         Table table = currentState.getDb("test").getTable("mv9");
         List<String> createTableStmt = Lists.newArrayList();
-        GlobalStateMgr.getDdlStmt(table, createTableStmt, null, null, false, true);
+        AstToStringBuilder.getDdlStmt(table, createTableStmt, null, null, false, true);
         Assert.assertEquals("CREATE MATERIALIZED VIEW `mv9` (`k1`, `k2`)\n" +
                 "DISTRIBUTED BY HASH(`k1`) BUCKETS 10 \n" +
                 "REFRESH MANUAL\n" +
@@ -126,7 +127,7 @@ public class ShowCreateMaterializedViewStmtTest {
         currentState.createMaterializedView((CreateMaterializedViewStatement) statementBase);
         Table table = currentState.getDb("test").getTable("mv10");
         List<String> createTableStmt = Lists.newArrayList();
-        GlobalStateMgr.getDdlStmt(table, createTableStmt, null, null, false, true);
+        AstToStringBuilder.getDdlStmt(table, createTableStmt, null, null, false, true);
         Assert.assertEquals("CREATE MATERIALIZED VIEW `mv10` (`c1`, `c2`)\n" +
                         "DISTRIBUTED BY HASH(`c1`) BUCKETS 10 \n" +
                         "REFRESH MANUAL\n" +
@@ -155,7 +156,7 @@ public class ShowCreateMaterializedViewStmtTest {
         currentState.createMaterializedView((CreateMaterializedViewStatement) statementBase);
         Table table = currentState.getDb("test").getTable("mv8");
         List<String> createTableStmt = Lists.newArrayList();
-        GlobalStateMgr.getDdlStmt(table, createTableStmt, null, null, false, true);
+        AstToStringBuilder.getDdlStmt(table, createTableStmt, null, null, false, true);
         Assert.assertEquals(createTableStmt.get(0),
                 "CREATE MATERIALIZED VIEW `mv8` (`l_orderkey`, `l_partkey`, `l_shipdate`)\n" +
                         "DISTRIBUTED BY HASH(`l_orderkey`) BUCKETS 10 \n" +
@@ -197,7 +198,7 @@ public class ShowCreateMaterializedViewStmtTest {
 
         MaterializedView mv = starRocksAssert.getMv(starRocksAssert.getCtx().getDatabase(), mvName);
         List<String> createTableStmt = Lists.newArrayList();
-        GlobalStateMgr.getDdlStmt(mv, createTableStmt, null, null, false, true);
+        AstToStringBuilder.getDdlStmt(mv, createTableStmt, null, null, false, true);
 
         Assertions.assertTrue(createTableStmt.get(0).contains(refreshClause), createTableStmt.get(0));
         Assertions.assertTrue(createTableStmt.get(0).contains(orderBy), createTableStmt.get(0));
