+++ fe/fe-core/src/test/java/com/starrocks/analysis/ShowCreateViewStmtTest.java
@@ -24,6 +24,7 @@ import com.starrocks.qe.ShowExecutor;
 import com.starrocks.qe.ShowResultSet;
 import com.starrocks.server.GlobalStateMgr;
 import com.starrocks.sql.analyzer.Analyzer;
+import com.starrocks.sql.analyzer.AstToStringBuilder;
 import com.starrocks.sql.ast.CreateTableStmt;
 import com.starrocks.sql.ast.CreateViewStmt;
 import com.starrocks.sql.ast.DescribeStmt;
@@ -189,7 +190,7 @@ public class ShowCreateViewStmtTest {
 
             List<Table> views = GlobalStateMgr.getCurrentState().getDb(createViewStmt.getDbName()).getViews();
             List<String> res = Lists.newArrayList();
-            GlobalStateMgr.getDdlStmt(createViewStmt.getDbName(), views.get(0), res,
+            AstToStringBuilder.getDdlStmt(createViewStmt.getDbName(), views.get(0), res,
                     null, null, false, false);
 
             Assert.assertEquals(testcase[2], res.get(0));
@@ -208,7 +209,7 @@ public class ShowCreateViewStmtTest {
 
         List<Table> views = GlobalStateMgr.getCurrentState().getDb(createViewStmt.getDbName()).getViews();
         List<String> res = Lists.newArrayList();
-        GlobalStateMgr.getDdlStmt(createViewStmt.getDbName(), views.get(0), res,
+        AstToStringBuilder.getDdlStmt(createViewStmt.getDbName(), views.get(0), res,
                 null, null, false, false);
         Assert.assertEquals("CREATE VIEW `test_view` (`k1` COMMENT \"dt\", `k2`, `v1`)\n" +
                 "COMMENT \"view comment\" AS SELECT `test`.`tbl1`.`k1`, `test`.`tbl1`.`k2`, `test`.`tbl1`.`v1`\n" +
@@ -278,7 +279,7 @@ public class ShowCreateViewStmtTest {
         List<Table> tables = GlobalStateMgr.getCurrentState().getDb("test").getTables();
         Table commentTest = tables.stream().filter(table -> table.getName().equals("comment_test")).findFirst().get();
         List<String> res = Lists.newArrayList();
-        GlobalStateMgr.getDdlStmt("test", commentTest, res,
+        AstToStringBuilder.getDdlStmt("test", commentTest, res,
                 null, null, false, false);
         StatementBase stmt = SqlParser.parse(res.get(0), connectContext.getSessionVariable()).get(0);
         Assert.assertTrue(stmt instanceof CreateTableStmt);
@@ -289,7 +290,7 @@ public class ShowCreateViewStmtTest {
         List<Table> tables = GlobalStateMgr.getCurrentState().getDb("test").getTables();
         Table storageTest = tables.stream().filter(table -> table.getName().equals("storage_test")).findFirst().get();
         List<String> res = Lists.newArrayList();
-        GlobalStateMgr.getDdlStmt("storage_test", storageTest, res,
+        AstToStringBuilder.getDdlStmt("storage_test", storageTest, res,
                 null, null, false, false);
         Assert.assertTrue(storageTest.isOlapTable() &&
                 ((OlapTable) storageTest).getStorageType() == COLUMN_WITH_ROW);
