+++ fe/fe-core/src/test/java/com/starrocks/catalog/CreateTableLikeTest.java
@@ -39,6 +39,7 @@ import com.starrocks.common.AnalysisException;
 import com.starrocks.common.ExceptionChecker;
 import com.starrocks.qe.ConnectContext;
 import com.starrocks.server.GlobalStateMgr;
+import com.starrocks.sql.analyzer.AstToStringBuilder;
 import com.starrocks.sql.ast.CreateDbStmt;
 import com.starrocks.sql.ast.CreateTableLikeStmt;
 import com.starrocks.sql.ast.CreateTableStmt;
@@ -81,9 +82,9 @@ public class CreateTableLikeTest {
 
     private static void checkTableEqual(Table newTable, Table existedTable) {
         List<String> newCreateTableStmt = Lists.newArrayList();
-        GlobalStateMgr.getDdlStmt(newTable, newCreateTableStmt, null, null, false, true /* hide password */);
+        AstToStringBuilder.getDdlStmt(newTable, newCreateTableStmt, null, null, false, true /* hide password */);
         List<String> existedTableStmt = Lists.newArrayList();
-        GlobalStateMgr.getDdlStmt(existedTable, existedTableStmt, null, null, false, true /* hide password */);
+        AstToStringBuilder.getDdlStmt(existedTable, existedTableStmt, null, null, false, true /* hide password */);
         Assert.assertEquals(newCreateTableStmt.get(0).replace(newTable.getName(), existedTable.getName()),
                 existedTableStmt.get(0));
     }
